Respect version constraints for cookbook dependencies
Move cookbook version dependency resolution out of nodes controller
Cookbook version constraints are handled in the RunList class and
accessible via the constrain_cookbooks on node objects.

Introduce CookbookVersionUnavailable error used when a requested
cookbook is not available to a node at -- either at all or at a
version that satisifies the specified constraint.