store the currently executing action on the provider
this isn't very pretty, but it allows the provider to choose what status
it will set on a resource based on the currently executing action.
Primary use for this is to choose whether to set the status of a service
resource to enabled/disabled or running/stopped based on whether the
provider is enabling it or starting it.