[CHEF-3557] temporary workaround for checksum cache race
File checksums are cached with the mtime of the file. When a file is
updated without updating the mtime (happens frequently on Mac OS X,
where mtime resolution is 1s), the checksum is invalid. We tend to see
this issue when we check, modify, then check again, so this change fixes
that by considering the cache invalid if the mtime as a float is equal
to the current time as an integer. This of course does nothing for the
case that you cache the value, change it, and then don't check it again
for longer than 1s.